Refactor installSqls command to install sqls and enable SQL language features
